fix: Long을 BigInteger로 변경한다
- JPA 결과 반환 시 BigInteger로 반환되기 때문에 BigInteger로 받은 후 long으로 변환한다

+++ BE/teamgu/src/main/java/com/teamgu/database/repository/UserChatRoomRepositorySupport.java
@@ -93,7 +93,7 @@ public class UserChatRoomRepositorySupport {
 						"select 0 as unread_message\r\n" + 
 						"from dual\r\n" + 
 						"limit 1";
-		List<Long> res = em.createNativeQuery(jpql)
+		List<BigInteger> res = em.createNativeQuery(jpql)
 							.setParameter("user_id", user_id)
 							.setParameter("room_id", room_id)
 							.getResultList();
